requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
hdx-python-api==6.2.6
pandas==2.0.3
slugify~=0.0.1
azure-functions
